+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 - Added the ability to whitelist only certain words in a pipeline
 - Fixed an issue with the PDF Parser not returning the first page
+- Added the ability to parse WebVTT file formats
+- Added `utils.listFiles` and `utils.walkdir` procedures